server/requirements.txt
flask_bcrypt==1.0.1
bcrypt==4.0.1
flask_cors==4.0.0
PyJWT==2.8.0
jwt==1.3.1
pymongo==4.6.0
python-dotenv==1.0.0
stripe==7.4.0